disable AgpVersionConsistencyTest
The AgpVersionConsistencyTest is loading two AGP versions to
check that we catch such errors. Unfortunately, the "older" plugin
which is 7.1.0 is using an internal Gradle API which has been
removed in 8.11.1 generating a different exception than the
expected one.

I suspect we will only be able to re-enable this test once we have
at least released an AGP version that does not use that API.

Bug: 382773586
Test: disabled a test.
Change-Id: I2094ee3cdde4aa13fa3d91f17493aa93df4a608c