GradleRule: More generic binary plugin API
and add support for old DSL/API.

Since the extension for core AGP plugin uses the new
DSL interfaces, there is no access to the old DSL
(normally only available through groovy). This is
problematic for the few tests that validates the old
variant API.

This change uses the same binary plugin mechanism, creating
a callback that uses the implementation class of the
DSL so that we can access legacy DSL/API.

Only supports Application for now but it's trivial to add
the other plugin types.

In addition with the introduction of this new plugin,
and more plugins coming soon (e.g. fused library),
this change renames part of the API to make it more
generic (it's not abount the component extension
anymore).

Bug: N/A
Test: migrating tests to new fixtures
Change-Id: Ia60ac770cbfaa531d064572690d2cf221f412c32