GradleRule: revamp build definition options
Move Gradle properties to the build definition.
This allows included builds to have their own.
Also made sure that the jdk properties are written
to the included builds as well.

Moved build name and root folder to the build
definition as well, as this is also relevant
to included builds.

More options are likely to move to the build
definition in the future.

Revamped the test location information as the
value returned by the original helper method
included the project (root build) name. This
is not convenient as we need the parent path
regularly and it made everything a bit awkward.

Bug: N/A
Test: existing tests
Change-Id: I3398bc867d5fdd86cbb459a5882499bfa5f657e7