GradleRule: support for getDefaultProguardFile
Since this method just returns a value there was no
link between the call and the result being passed to
another method.

this change introduces the first modified type: MethodReturnedFile
which extends File. When the BuildWriter sees such a type to
write (as another method parameter), it knows how to reconstruct
the original call to pass to the method.

This also updates R8TaskTest which uses getDefaultProguardFile
and refactor some of its part of make the configuration a bit
simpler to follow. Part of this removes the executor and
app getters that just aren't needed, practical or safe to use
due to them calling rule.build.

Bug: 384016091
Test: Updated R8TaskTest which uses this pattern.
Change-Id: I98ffa249f8833b51cfe5e50607aa3eee46079ead