GradleRule: improve JarBuilder API
Instead of set methods, this is now only
add method that allows building a larger Jar
with many different classes and text files from
different sources (actual classes, empty classes,
classes with empty methods).

This also changes how adding classes with empty
method works as separating the method name and
the descriptor with : does not make a lot of sense
when there is the parenthesis.

Bug: n/A
Test: updated existing tests
Change-Id: I6b3fbed15dffa1595381d7b729b785a66d515731