GradleRule: Convert more kotlin tests
UpgradeKotlinGradlePluginApiTest does not actually do anything at
the moment because both the built-in Kotlin support and KGP have
the same version. We need to fix this.

This change also adds support for adding arbitrary dependencies
to the buildscript classpath of a project. This is used in this
test to add a dependency on KGP.

Bug: N/A
Test: converting tests to new fixtures
Change-Id: If50ae37a7c569635b8efa1a56f8c758da366c7b6