374282903: NewApi does not catch use of InstantSource
Also fixes a bug where we merge multiple supported method
lists; if one file lists an entire class as backported,
we should drop any individual member entries from the
other file. An example of this is java.util.Objects which
has a handful of supported methods in the default D8
backport list, and the entire class included with library
desugaring.

Fixes: 374282903
Test: Unit test included (and expanded)
Change-Id: I4757ae21648ecb08a02350993e63dfc4e6c760c9