Fix "is partially supported" desugaring method
The signature files from D8 are missing constructors, so until
they do, lint has a workaround where it considers a constructor
supported if there are other methods in the class that are
supported. (This heuristic would be better if it could limit this
to instance methods, but the signture files don't carry that
information).

This CL broadens the unit test for desugared element lookup
to not only positively check for all elements listed in the desugared
files, but it also uses the API check database file to also ensure
that all other valid APIs are *not* matched. And this revealed
one other corner case bug: constructors. The check to see if
a constructor is implicitly supported (by looking for other methods)
was doing a simple string prefix check, which would also match
other classes that have the same prefix; it needs to make sure
that it only matches members for the same class.

Test: Unit test included
Bug: Not filed
Change-Id: Id8c14b4a4a6ee1a81d039aad1f2c3800a4a1e861