Pass ExecutorService to R8
so that we reuse a shared thread pool across R8 invocations.

The thread pool and its size match R8's default settings, and we aim to
keep them in sync. The only difference this change is making is that
instead of R8 creating a new thread pool per invocation, AGP will use
the shared thread pool for all R8 invocations.

Bug: 375394051
Test: Existing tests
Change-Id: I7fdbad603e50462921ad4683ae277f60bd613a6a