Make comparison well-defined in AndroidVersion
Define comparison using comparators to make it more understandable, and
make it follow the contract of Comparable.  The previous behavior of
having a special case for when isBaseExtension is set on both operands
violates the transitivity requirement for comparators.

Delete compareTo(int, String): it's not obvious what this means in the
presence of extension levels. Provide isAtLeast(int, String), which more
clearly handles what it was being used for.

Delete isGreaterOrEqualThan(int, int), which was not used, and had
unclear handling of base extension levels. (Note that the assertions
made by its tests differed depending on isBaseExtension.) These
compaisons can be made using AndroidVersion.compareTo(AndroidVersion).

Deprecate isGreaterOrEqualThan(int) in favor of isAtLeast(int), which is
more grammatical and readable.

Bug: 389973543
Test: AndroidVersionTest
Change-Id: I16db02a20152d9368f4a995262c3e05f11d67f66