Set Compose plugin's sourceInformation to true for release builds
Previously, the flag was only set to true for debug builds. In line with
KT-74415, the flag should always be true by default.

In a subsequent change, AGP will defer to the Compose Compiler Gradle
plugin to set this value so that it will be configurable via the Compose
Compiler Gradle plugin's DSL.

Bug: 362780328
Test: ComposePluginOptionsTest
Change-Id: I0b3d1020dabcf38b6a7a2ca1d36f273bacb16b3e