Update issue doc generator to handle hints
Also, for examples extracted from unit tests, it
now includes icon images inline, and for some binary
files (stored as base64 gzipped in the test) they're
now included as links.

Also adds more caching to make it faster when run
repeatedly.

And fixes a typo in an issue option explanation.

Test: Manual (inspecting the diffs in the docs update,
    https://github.com/googlesamples/android-custom-lint-rules/commit/65850ee3d7fe594293df343666f94e33b49f1cb2)
Bug: None
Change-Id: I4d3ceb2468a740d56b7d7fd040b6763cf46af2eb