Make the 16 KB alignment detector work on the fly
This CL allows the 16 KB alignment lint check to run on the fly; when
you're editing build.gradle or TOML files, the dependencies are checked
to JNI alignment issues.

To avoid doing expensive I/O repeatedly, it has a pre-computed table of
all the possible AndroidX libraries that have to be checked (the rest
are assumed to be fine), as well as a cache of group+artifact+version
libraries it has already checked.

Also fixes
395836302: ConcurrentModificationException in Aligned16KB lint check

Test: Tests included
Fixes: 395836302
Change-Id: I651ff5fa7a77e728ff1da8ac05f64407cc0326f8