Work around DaemonDisappearedException in AllDslMethodsTest
This change adds the following JVM arguments to AllDslMethodsTest to
work around an unknown JVM bug which is causing
DaemonDisappearedException:

  -XX:CompileCommand=exclude,org.jetbrains.kotlin.serialization.deserialization.TypeDeserializer::simpleType
  -XX:CompileCommand=exclude,org.jetbrains.kotlin.serialization.deserialization.TypeDeserializer::toAttributes

See the bug below for more context.

Note: I'm not able to reproduce this issue with --runs_per_test so I'll
continue to monitor the flakiness rate after this change is merged.

Bug: 382709449
Test: AllDslMethodsTest
Change-Id: I037defb827b7082f68688acbfaf2df1628640e83