Update com_google_protobuf and grpc
grpc is updated to 1.69.0, and is pulled from the Bazel Central Registry
(BCR). This requires making some changes to how abstract sockets are
used, since the grpc version in external/grpc-grpc relied on custom
patches to support the feature. The BCR version still needed patches to
migrate to Bazel platforms.

protobuf is updated to 29.2 and needed a patch to disable micro protos
(upb). There is an upb aspect that has an error when bazel runs with
"-c opt".

zlib is pulled from BCR, and patched to replace a genrule with skylib's
copy_file rule. Skylib could not be added to zlib's dependencies since
that would require a new BCR release where MODULE.bazel is specified.

Other dependencies are updated as they are needed by grpc and protobuf.

Presubmit-test: default
Change-Id: Idade066e828dea92e5d7634ba35e651ce9c565f7