Use rules_android from BCR
Bazel no longer bundles Android rules, so using rules_android is
required for updating to Bazel 8.

The ANDROID_HOME environment variable was moved out of bazelrc files
to fix androidx-vitals integration test. The AndroidX team set
ANDROID_HOME in their own build, so we need to use that value.

The aosp_studio_linux build does not have an Android SDK available, so
the script stubs one to make rules_android initalize. The build passes
because no Android targets are built on the aosp_studio_linux target.

The validate_aar check is disabled on Windows due to paths on RBE being
broken, but this is fixed in Bazel 8.

Bug: 371633138
Change-Id: I8d7d1645e018ff14d352800e9dd3f7c1f9c182ec