GradleRule: rework chained proxy.
the previous implementation was inconsistent
in different places. Between parentChain,
ChainedContentHolder and the manual
implementation of collection proxy, they all
handled chaining differently.

This change cleans this up. Each proxy object
now has its own ContentHolder which record the
API calls. (no more ContentHolder facade
that just delegate the recording to a different
instance.)

Instead of a content holder recording, at
the event time, the chain of the parents,
the events are now always recording locally
with not chain information. Only at write
time does the parent will pass its name
to the chained holder so that it can
write the right chain. This simplifies
many calls and localize the chain handling
to the write method only.

This introduces a new EventType when a chain
is created which allows a content holder to
call the chained holder at write time.

This also removes the name on the holder
as it's not necessary (the names are attached
to the event).

Finally, this also clean up some code by removing
from the content holder the creation of proxies.
This class' responsibility should only be of
recording event and sending them to the writer.

Bug: N/A
Test: added more unit tests
Change-Id: I8363611d0846e04263aa1c785fceb3c9070c2cd9