GradleRule: support injected DEFAULT_COMPILE_SDK_VERSION
GradleTestProject has a DEFAULT_COMPILE_SDK_VERSION constant
that can be fed via an env var ("CUSTOM_COMPILE_SDK"), so this
needed to be ported over to GradleRule.

This introduces a field on GradleBuildDefinition, as well as one
on its Companion so that it's easily accessible from the DSL
when building a project.

This also documents and mark as deprecated the standalone
constant DEFAULT_COMPILE_SDK_VERSION so that it's not used
directly. This is the base value, but isn't overiden by
the env var.

Finally, fixed many usages in the "new" tests (ie the tests
migrated to GradleRule) since they were using the wrong
versions.

Bug: N/A
Test: updated tests
Change-Id: I979bf1164fb53cfffcfe1797cba3a2754ef56198