removed API to set multiple providers in Artifacts
Public APIs do not have an API to set mulitple providers in a
single call. For instance, transform requires to combine all
the inputs into a single output file or directory.

Therefore PropertyAdapter do not need an API like
set(Provider<List<T>>) and it is enough to just have set(Provider<T>>.

As it turned out, the OutputRequestImpl was doing something
incorrect because of this API. It forced the combined folder
obtained from the Transform to be stored into a new ListProperty
to comply with the API. This broke the task dependency as the
ListProperty was not tagged as an output of the underlyng task.

Futher simplification, enhancements of the kotlindoc followed.

Test: Existing
Bug: N/A
Change-Id: I63ca060b185400d484b7180ea8e2afd5934c1cfa