[GMD] added an error for GMD vendor string
Added an error when a developer attempts to specify a page size  as
part of the system image source. Although technically the page alignment
ends up being part of the vendor string in the context of System image
hashes, the actual part of the Managed Device DSL is meant to handle
page alignment is the pageAlignment Property.

The reason behind this is 2 fold: 1. the recommended system image page
size is dependant on the api selected. I.e. one may want to select a
source and use the recommended page size, and we can't tell if
"google-apis" means just "a google image with the recommended page
size" or explicitly "a google image with a 4kb page size". And 2.
This makes the behavior of the DSL -> system image easier, and makes
suggesting recommendations of system image selection easier as well.

Bug: 382551039
Test: Updated/added
Change-Id: I04adef89cc56a845865fd244f86d23ca847ea92c