Migrate to non-deprecated methods and simplify
AndroidVersion.getApiStringWithExtension already implements the base
extension level suppression performed by this code; just call it instead
of the deprecated getApiString.

This does change behavior in the case of an AndroidVersion with a
codename and a non-base extension level. For these versions, we don't
want to include the extension level. We include the extension level when
it is greater than the version at release (the "base"). The preview has
an extension level that is generally less than or equal to the level
that it will be released with.

Bug: 393164974
Test: DetailsTypePathsTest
Change-Id: Icff6fe69d4054964f1071e1d10b7e0b1894996e1