Stabilize HostTest, DeviceTest, HasHostTest and HasDeviceTest
Interfaces are stable but some of their methods are still in
incubating mode.
getBuildConfigFields was changed everywhere to return a
nullable MapProperty to reflect the optional nature of the
build config fields which are an optional feature.
Null objects will not be be returned until AGP 10. In the meantime,
a warning (already present) is issued when using these methods
when the feature is turned off.

Test: Existing
Fixes: 357682409
Change-Id: Id6117706bd322f55d7e0e9042ab6a9a28f8b9db5