[GMD] support testedAbi dsl
This adds support for the testedAbi field to Gradle Managed Devices.
This field allows for specifying the abi of the apk that will be
installed for tests. At present the abi would match the abi of the
system image for the given managed device.

As the system image would change based on the developers environment,
(such as using an ARM image on a mac m1) the native code being run on
a specific GMD would depend on the machine running it. As this is
against the spirit of GMD, where the tests should be the same as
possible even when run on different devices, we added the ability to
target the abi for the apk.

If the tested abi is left unspecified, this results in the GMD following
the old abi selection path, but will default to arm64 (as
this is the abi most real world devices use) in 9.0.

Bug:382551039
Test: Added

Change-Id: Ib1133f3a2c6ab32ec95e8ed7e5c031570914c00c