First iteration of test suites variant API.
TestSuite/Builder instances are created so far, and accessible
through the existing beforeVariants and onVariants callbacks.

There is an instance of TestSuiteBuilder for each test suite that
applies to a specific variant. Since the compilation and test
task will be specific to the tested variant, each TestSuiteBuilder
will be able to be customized independently despite they all
point to the same test suite definition (and sources).

Similarly, the code is not able to disambiguate the Host tests
versus the Device tests so far.

Eventually a Test Task is created for each TestSuite instance
but that Task is not doing much so far but at least validates
the Variant objects population

Test: added unit tests and integration tests
Bug: N/A

Change-Id: I000b2f0ed27f312da6ef6907756fcb89530a0058