Extend LintJarApiMigration to rewrite member-level API changes
Previously, it depends on typealias hints that were left for backward
compatibiilty purpose. Based on the type mapping (from typealias),
it performed class name rewriting all over the type appearances.

In reality, rewriting should have been two steps. For example,
KaResolver's `KtElement.resolveCall()` appears as
  `KtAnalysisApiSession.resolveCall()`
after being mixed in to the session.
With the current type mapping, it is rewritten to:
  `KaSession.resolveCall()`
which worked so far because there is _another_ deprecated API that
simply delegates the call to the right one:
  `KaSession.resolveToCall()`
In addition to typealias, such bridges are also gone in 2.1.20.

To that end, this commit edits ExtractMigrationTable to
1) visit members, in addition to typealias;
2) check if they are annotated with @Deprecated(...);
3) retrieve `replaceWith` hint in the annotation; and
4) build API mapping, along with simplified signature (*)

(*) to handle API name clash, e.g.,
KtExpression.getKtType() -> getExpressionType()
  v.s.
KtTypeReference.getKtType() -> getType()
(Unit test is added for this specific case.)

Then, this commit extends LintJarApiMigration to apply API mapping
in addition to type mapping.

Bug: 392185652
Bug: 393478604
Test: LintJarApiMigrationTest
Change-Id: Icb214afd2ca5f44bfb32eef36587f0008ece86f4