Remove IJ Platform owners
These are not needed anymore due to the new Global-Approval label,
which can be used to override owners approval for straightforward
large-scale changes. The Global-Approval permission is already
granted to the right people on the platform team.

This change aids the "Suggest Owners" feature in Gerrit, and it
prevents accidental owners approval for files outside the normal scope
of platform team (e.g. see Change I84b956f40).

Change-Id: I8aa8e22f89f173d9d3cd50aecad4265597c9a374