Add a DslMarker for the GradleRule DSL
This will avoid misplacing calls due to
implicit receivers. For instance,
before this, it would be possible to place
the creation of a project inside the configuration
of another.

In fact, this showed one test written incorrectly,
trying to get included builds inside an android app
module.

This change also required a different handling for
DEFAULT_COMPILE_SDK_VERSION as it does not really
work as a field on GradleBuildDefinition. Instead,
all usages were moved to the value on the companion
class.

Bug: N/A
Test: existing tests.
Change-Id: I81d27cb676f45c85b7ff3cd26a6f38cd9c552508