Implement basic on-device rendering
Implement code to enable the agent to render Layout Inspector view
bounds on the device.

This is implemented so that the source of truth remains on Studio,
therefore the inspector will detect touch events on the device, then
send them to Studio, which will reply with the view bounds of the
view that needs to be rendered.

The on-device rendering is intentionally "dumb", in the sense that
it only know that it needs to draw rectangles at certain positions.
This makes sure that the source of truth is in Studio and that
studio-side rendering is consistent with on-device rendering.

Test: OnDeviceRenderingViewModelTest
Test: OverlayViewTest
Test: ViewLayoutInspectorTest
Bug: 398201661
Change-Id: I4b9b828e0a5325bf5c395388608f0a32eed29498