Break getSdkManager() into two parts
getRepoManager() instantiates the RepoManager, but does not load any
local packages; it will complete quickly.

getRepoManagerAndLoadSychronously() is a direct replacement for the old
getSdkManager(), which both instantiates the RepoManager and calls
loadSynchronously() on it.

This makes it clearer that the method is potentially slow and should not
be called on the UI thread.

Also, some places get the RepoManager and immediately call
loadSynchronously on it with custom arguments
(e.g. SdkComponentsStepController.onPathUpdated),
but those custom arguments are irrelevant since the loading already
happened with the default arguments within getSdkManager().

In other cases, loadSynchronously is immediately called with zero
timeout, discarding the loading that was just done (e.g. AGP's
DefaultSdkLoader).

Bug: 339641500
Test: existing
Change-Id: Ifd4ead4b508f6771f113925d066f6e7007266b05