Add generic analysis for properties with simple "join" structure
This change implements a generic, compositional analysis for any
property that follows the simple structure where the analysis of the
whole is a join over the analyses of the parts, regardless of whether
the results are combined from different branches or sequentially along
the same path.

To implement an analysis for a concrete property, the user will supply
a lattice (describing monomorphic results), and specify where these
(monomorphic) results are introduced. The result of polymorphic
functions are automatically polymorphic.

At the moment, it only exposes enough API for implementing analyses
where effects are introduced as method annotations. The API will be
extended to be more flexible in the future.

Bug: n/a
Test: new for some utils; higher-level tests deferred to future CLs
Change-Id: Ib4c9f34cce932d8fa8042e9414cd6af3c6cd2c61