Subjects for classes and resources
This introduces class and resource specific subjects, that
are meant to test the content of AAR, APK, and internal jars.

In a later commit, JarSubject will be changed to use this directly
instead of having its own implementation. Similarly, AarSubject
and ApkSubject will just delegate class and resource testing to
these objects.

There are different implementations for APK and AAR since
code is handled differenly (inner jars vs dex files), and java
resources is alos in a different location (directly in APK vs in
inner jars)

Bug: N/A
Test: unit tests only for now.
Change-Id: I4b593781631c20825b54304a398cdab6d32015af