Implement on-device rendering of hovered and visible nodes
This change adds support for rendering of hovered nodes and visible
rectangles.

Visible rectangles are all the rectangles visible on screen, Studio
sends them over each time there is a model change.

Hovered rectangles are the rectangles currently being hovered. The
on-device rendering detects the hover coordinates, sends them to
Studio, which checks the model and responds with the draw instructions
for the view node at those coordinates.

Test: OnDeviceRenderingViewModelTest
Test: OverlayViewTest
Bug: 398201661
Change-Id: Id873ce51b9490196ec0a5be0be1926a4cfac0f6e