Fix missing version exception from GoogleMavenRepository
The findCompileDependencies method on GoogleMavenRepository
reads POM files and returns the set of Dependency instances
it finds for the compile scope.

However, for some POMs, like androidx.activity:activity-compose,
this method throws exception, because the code requires every
<dependency> to include a version definition. But this is
not always the case; in the above POM for example, we have
the following:

    <dependency>
      <groupId>org.jetbrains.kotlin</groupId>
      <artifactId>kotlin-stdlib</artifactId>
      <scope>runtime</scope>
    </dependency>

This means attempting to read dependencies from this library
results in an exception instead.

However, the actual dependency is defined earlier in the file
under a <dependencyManagement> node. So, we can actually resolve
this version. (Furthermore, this dependency actually had a
runtime scope, and we're looking for compile time scope, so
we shouldn't bother checking dependencies that we're going to
ignore anyway).

Bug: Not filed
Test: New unit test added
Change-Id: I577c7675d27552ee54f438d480793f8818f1a96e