Revamped Task subject
replaces the old GradleTaskSubject mostly to move
it to Kotlin and to provide better error messages.

Also direct (and only) access to it via GradleBuildResult

This is a first step toward removing direct access
to the executed task list in order to remove
usages of doesNotContain which is not safe.

Bug: N/A
Test: None for now. will add unit tests later
Change-Id: I209b1e36acf97707ae2f090143cee4dc52aea70e