Improve handling of non-gmaven dependencies
We've had a lint check which checks non-gmaven dependencies
for a while, "NewerVersionAvailable", but since it was quite
expensive (hitting the network each time for every dependency,
running a query on maven central json-based search mechanism),
it was off by default. We also had a more recent mechanism
to look up Gradle plugins on Gradle's plugin portal. Both of
these had a TODO to add some caching as well.

This CL switches this over to reading maven-metadata from
maven central directly instead of invoking a search, and wrapping
this with a cache (similar to what is already done for gmaven),
as well as unifying some of the code around this.

Because of these improvements, we'll also turn this check on
by default.

Second, it improves the way we suggest versions. Until now,
we'd simply consult the remote repositories asking for the latest
stable and the latest preview version. If the current version
was a preview, we'd update to the latest preview, and otherwise
latest stable.

Instead, we now pass in the *current* version to the search
mechanism, and then we pass back a data object with various
version suggestions:
(1) Latest preview version
(2) Latest stable version
(3) Latest *patch* version
(4) Suggested version

The suggested version to use is the one we'll suggest in lint.
This is generally going to be the latest *stable* version that
is higher than the current version (unless only previews are
newer, and the current version is also a preview). This
means we won't promote you from some old preview (say 7.3-alpha03)
to the current preview (8.10.0-alpha01), we'll suggest 8.9.0.

There's also the patch version, which is a more recent patch
version of the same major+minor version you're already using.
For AGP specifically, we can offer to directly switch to this
patch version, and for higher versions, we'll suggest the AGP
upgrade assistant.

There was also some old hardcoded upgrades for specific
versions in there; removed this (the versions are long obsolete,
and these checks were there because we weren't looking them
up automatically; now we do.)

Bug: Not filed
Test: Unit tests included
Change-Id: I105aead9eb79fd127efd0c76654280e0b195d3b4