[fused lib] Support resolving transitive dependency constraints
This change adds a new resolvable transitive configuration for the purpose of resolving version constraints of dependencies. This is needed to support adding platform/bom dependencies. The `includeTransitiveResolved` configuration is only used for version aligning when resolving non-transitive dependencies.

A disadvantage of this approach is that transitive dependencies do now get resolved, which means that if there are databinding dependencies present in dependencies, the fused library validation may not detect them before configuration resolution fails.

Bug: 399879853
Test: FusedLibraryClassesVerificationTest
Change-Id: Ie9491cf0a33d0b3ea77f9c1123942c5e6ed00262