Revert^2 "Update com_google_protobuf and grpc"
860bb620c432e289602f0d70a0793b08e54cdf17

Disable the supports_dynamic_linker feature on MacOS. This resolves the
thread-local linking error. The diff between this and the original
commit is in bazel/platform_specific.bazelrc. This results in the
artifacts on MacOS being a few more MB in size.

See https://github.com/bazelbuild/bazel/issues/4341#issuecomment-758361769

Presubmit-test: default
Change-Id: Iadb5af26a8c3027317a8716d740783707f089f07