When waiting for a job do it outside the critical section
Fixes following issues:
When we identify that the same transition is already in progress
 we should wait on that job to complete outside the critical session.

Also, if start fails, and then we trigger restart, it should retry
a start. It can be argued that we should instead throw
an `IllegalStateException`, but for now let's just start
a server as a `restart` caller may not be aware of the current
server state (the server state could change any moment).

Bug: 378111505
Test: added
Change-Id: I68acd69f85850403914576e3c0aca3d0cfca1cf7