Initial wiring of custom junit test engines.
When declaring a new test suite, one can now use the includeEngine
DSL declaration to indicate which test engines should be used to
execute the test suite tests.

The junit test engine must resolved using the test suite `implementation`
configuration. So for example, if you create a test suite called 'first',
and its dependencies can be declared using the dependencies block under
the test suite block. Such dependencies are only added to the test suite
being configured. At this point, only implementation, compileOnly and
runtimeOnly are supported.

Users must at least add the junit engine dependency so it can be loaded
correctly. There are also a few other system depdencies like the junit
platform engine that must be added manually so far. In a subsequent CL,
we should consider adding those automatically.

Input parameters declared through the DSL are passed to the junit
engine as elements in a json file identified by an environment variable
passed to the junit engine process.

The junit engine can retrieve such inputs using the json file.
Some rudimentary logging facility and streaming facility was added but
will be improved in a subsequent CL.

Sync in Studio is not functional yet, as ModelBuilder expects all test
suites to have a JAVAC task which is not the case at this point. This
will be addressed in a subsequent CL.

Test: Added unit and integration tests
Bug: N/A
Change-Id: If0a1d817f4350f19faee9e4e3156bf4cbb3760f8