Make GoogleMavenArtifactId enum names more systematic
For old support libraries: include the last element of the group
unless it duplicates the first element of the name.  For androidx
libraries: include an ANDROIDX_ prefix.  Convert kebab-case and
camelCase to SHOUTY_SNAKE_CASE.  Do not try to maintain direct
correspondence between old and androidx names (e.g. do not preserve
-v7 suffixes in androidx equivalents).

This is in preparation for importing a larger set of artifact ids to
support androidx migration data directly.

Bug: none filed
Test: existing
Change-Id: I1c7a3aeb9c697961eebead5252871129fffbc681