Add OverlayViews if they were removed by the framework
A root view can be missing the OverlayView if the root view children
were all removed, but the root view itself was not. This can happen
in some cases, for example if the Activity is relaunched.

This change makes sure the OverlayView is actually present, each time
LayoutInspector tries to interact with it.

Test: OnDeviceRenderingViewModelTest
Bug: 402729631
Change-Id: I61043cb4e741fd9e3db897719edab09bee3559b9