Add a utility to GoogleMavenArtifactId
A number of places in the system still want to find the equivalent
androidx library for a given (old) support library.  Wrap the
string-based operation, and include a test to verify that all the
artifacts in the map are represented as enum values in
GoogleMavenArtifactId.  (This adds a test module dependency on
sdk-common for the sdk.common tests, which is probably less bad than
expanding the visibility of ANDROIDX_COORDINATE_MAPPING given that
the eventual aim is to remove ANDROIDX_COORDINATE_MAPPING.)

Bug: 404519855
Test: added
Change-Id: I3468e1416afb2084316065c1ec78668e3f9d2695