Update kmp androidTest output location
This came as tests were migrated to use the right
API to test AAR/APKs, and it was noticed that android
Test APK for KMP libraries where put in a directory
with a /main/ segment. This was due to them not being
variant aware.

This change removes the main segment for all cases as it's
not necessary and is inconsistent with the other paths.

Bug: N/A
Test: updated tests
Change-Id: I4051f19f868666c408dd7512117786e0b4127819