Improve lint handling of TOML and properties files
Some Gradle files, such as version catalogs, and gradle.properties and
gradle-wrapper.properties, are shared between all Gradle projects, and
we want lint to flag problems in these files.

However, they aren't actually part of any *specific* projects you run
lint against. E.g. in a library that has an app module and a lib module,
which lint report do you expect to see these errors for? Initially, they
didn't appear anywhere, because again, these files aren't part of either
project. Then we started including for example
../gradle/libs.versions.toml, but that meant that the errors would be
duplicated in each project. So then we started trying to designate a
single project to be the "holder" project.

However, there are cases where this wasn't working well, and *that's*
what this CL is fixing.

First, it introduces a property on the Project which indicates whether a
project is the holder project for these files. In the IDE, where we also
have access to the real root project and can pass it to lint, we use
this mechanism.

From Gradle, we're never told about the root project, we're just invoked
on each sub project. In that case we use the previous heuristic, but
it's been improved a bit. It now has a more accurate way to find the
root Gradle project, and it scans the settings file to find the first
included project and associates the project with that.

Bug: Not filed
Test: Existing (and some new unit tests)
Change-Id: Ie59edfee91b9e11cba1d53e133017021ce5b6f64