Move kotlinAndroidProjectExtension to BuiltInKotlinServices
so that the handling of built-in Kotlin is consolidated in one class.

Note: This change is refactoring only, the only behavior change is
that screenshot tests now require the Kotlin Gradle plugin to be
applied (see the other change in the same topic). In the long run,
we don't need the Kotlin Gradle plugin anymore, so this is only
temporary.

Bug: 341765853
Test: Existing tests (refactoring only)
Change-Id: Ia8251d061684c894c2eac7c5e8cff686cebf315a