Replace ZipCentralDirectory with ZipFlinger.
Original impulse was to remove ZipCentralDirectory as it implements zip
operations that are offered by ZipFlinger classes.

Apart from this redundancy the class exposed API allowing creation of
zip files without their payload (as a way to lower the storage impact).

The new ZipSnapshotRepository allows taking snapshots of zip files,
which can be used for tracking incremental changes of.ap_ or .jar like
files.

Bug: 149571216
Test: Existing
Change-Id: I49fd03d90c2fc4065ff2139555ce32e896c74257