adblib: remove proxy info from server
Removes `JdwpProcessDebuggerProxyInfo` from the
process inventory server, proto definition, etc.

This is currently not needed, as this only enables
debugging processes when multiple instances
of Android Studio run concurrently. This was
never supported, and this will make follow CLs
more complex if we need to maintain this.

This can be re-introduced later if needed.

Bug: 333388357
Test: Existing
Change-Id: I6434c0d1d6714be399646f7c7818d42892f7b06f