Fix manifest testing.
the way the old fixture worked was problematic in several
ways:
- too lenient: missing manifest did not error out and just
  returned an empty string which could be problematic in
  "doesNotContain" tests. (this commit encountered one
  such test)
- testing manifest by strings is problematic anyway and
  we want to be able to test specific node content to
  avoid using doesNotContain

This commit introduces a new node-based API, only available
on APK for now. It allows drilling into nodes and checking
attributes and sub-nodes. We'll have to add it to AAR
subject by parsing the XML file.

All existing tests using the old ApkSubject were converted
to this new one.

We may want to consider getting rid of the pure text based
API in the future.

Bug: N/A
Test: fixing tests only
Change-Id: I09b950dc86c4808f05b73c9c10eac2997dfb2fac