Improve handling of implicit lambda receiver locations
In KTS files some DSL methods have a receiver of
type KotlinUFunctionCallExpression$KotlinUImplicitLambdaReceiver
and these did not have correct location computation.

This CL fixes this. Unfortunately, this behavior doesn't
get triggered in the CLI version of KTS, so the real test
for this is in AndroidLintGradleTest.testCompileSdkLocation
in the IDE lint module.

It also adds support for rootProject references in build
script apply blocks.

Bug: Not filed
Test: New + Existing + AndroidLintGradleTest.testCompileSdkLocation
Change-Id: Ic2634f6cbfbeeece74676c114922c01179a2cb0d