adblib: introduce screen recording api
Add support for invoking the `screenrecord` command
as an extension function of `AdbDeviceServices`.

Calling `screenrecord` on a device is a common operation
for tools (e.g. Android Studio), and it is actually somewhat
tricky to get right because the `screenrecord` command
relies on processing `Ctrl-C` from its `stdin` as the way
for the caller to "signal" the screen recording operation
should stop.

This behavior makes sense when the command is used
by an end-user in a shell terminal, as the user ends the
recording by pressing `Ctrl-C`, however it is non-trivial
to make this work for tools, so this changes introduces
a "wrapper" for that purpose.

Background: when `screenrecord` receives the `Ctrl-C`
signal, it finalizes the video recording file on the
device storage (i.e. it write the `mp4` file metedata and
flushes the last buffered frames), which means the
`mp4` file is valid only if `Ctrl-C` has been processed.
This implies that force killing the `screenrecord` process
instead of sending the `Ctrl-C` single results in an
invalid/corrupted video recording file.

This change adds a `AdbDeviceServices.screenRecord`
function that simulates the behavior of calling `screenrecord`
from a shell terminal, and give the caller control of when
to send the `Ctrl-C` signal through a `Deferred<Unit>`
parameter.

This change also add a few helper functions to
`ShellCommand` to allow escaping path name
for running shell commands on a device.

This change also fixes the way `FakeAdbServer` processes
shell terminal sessions to correctly support the `exit`
shell command.

Bug: n/a
Test: Included
Change-Id: I1d8de87d08ed3d7d635fd0f56123c8ba52adcc06