408498803: Trap 404's for dependency lookup when allowCache=false
When lint runs, it goes via a network cache when looking up
the latest versions of dependencies such that we don't hit the
network every time you're typing a build.gradle or version catalog
file.

However, the same lookup utility also lets you ask for a fresh result;
this is for example used by StudioBot's version lookup facility.
In this codepath we weren't trapping unexpected 404's.

Fixes: 408498803
Test: Existing
Change-Id: I182ba5b6dd447e371bd3b8ee92d232e9b43d48e9