Change platform attr in project.xml
Removed Project.getPlatform(). It is not clear what this function
returns (or will return in the future), particularly for Lint in the IDE
or Lint via Gradle. For project.xml, it was most likely returning one of
a limited set of "conventional" source set names ("ios", "common", etc.)
but these seem to have come from the Kotlin Multiplatform Gradle plugin,
where source set names should not be relied on (these initial source
sets are provided for convenience, but can be modified, and new source
sets can be added). We do not want to commit to this API forever, so I
am removing it because we can still achieve our goal without it, and the
`TargetPlatform` can be accessed by Lint checks via the AA (see below).

In the project.xml representation, changed the `platform` attribute to
`kotlinPlatforms` to make it more clear what it is (although it will
never be completely clear, as targets, platform(s), target platforms,
etc., are often used interchangeably). It takes a `TargetPlatform`.
Aka, a set of simple platforms, a set of Kotlin platforms, a set of
component platforms. This has a form like:

"Native [general]/JVM [1.8]"

and is deserialized by:

deserializeTargetPlatformByComponentPlatforms()

which is provided by Kotlin compiler sources, and looks to have some
backward compatibility effort, although we may want to come up with our
own approach for Gradle, where backward compatibility must be handled
very carefully. The `targetPlatform` can be accessed by Lint checks via:

analyze (ktElement) { useSiteModule.targetPlatform }

ManualProject is changed from private to internal (for now) so that
UastEnvironment can access the kotlinPlatforms property. For other
Project subclasses, we continue to just rely on source set name
conventions (but I added a TODO because we should improve this).

Bug: 406142870
Test: Updated ProjectInitializerTest
Change-Id: I04569bbd508d5031cc49cecea7377e93df15036d