Introduce GoogleMavenRepositoryV2 interface and it's skeleton implementation.
Currently, the new functions in "GoogleMavenRepositoryV2" return sample data but this is part of a wider effort to refactor GoogleMavenRepository to hide the implementation details from the callers. Next, we will introduce the implementations for the methods and enable it for Canary for a phased rollout.

More details in go/studio-template-dependency-impl

Bug: b/407727269
Test: Unit tests and WAI screenshots in comments
Change-Id: I0bccd552d124caa4f0506d608167fd32c4cef867