adblib: improve `screenrecord` error message
Improve exception message collected when running
the `screenrecord` command: in addition to the
exit code from the command, we now also try to
collect the error message so that we can report
to the user.

A common use case is when the screen of a physical
device is locked, `screenrecord` outputs specific
error message ("ERROR: INVALID_LAYER_STACK,
please check your display state.") that can be useful
to the user to understand why a screen recording
can fail.

This change also improves unit test coverage wrt the
`AdbScreenRecordException`.

Bug: n/a
Test: Included
Change-Id: Id6bcf1c5f9f47ce4acc8fc7800316b671b12648d