410936202: Lint warning for 16K doesn't show up in "Inspect Code" results
There are two related complications here:

First, the lint check checks the full (transitive) dependency graph
of libraries; in some cases, those errors are not explicitly mentioned
in a build file, so lint doesn't know where to "attach" the error
message, so in those cases it points directly to the unaligned shared
library file. However, those files are not inside the project directory
being analyzed. That's not a problem from Gradle etc -- but inside the
IDE, inspection results are automatically filtered to be inside the
scope -- there doesn't seem to be a way to have results from outside
the project.

This was exarcerbated by another issue in the IDE: In the IDE there
are two different implementations: one which runs on the fly inside the
TOML file and looks up each library (which works already), and running
lint in batch mode; this is the same code path as is taken from Gradle.
But here we fail to look up the corresponding library from the model
because in the IDE there's a separate wrapper created by the sync model
to contain root files like the TOML file, but there are no dependencies.
So we encounter the TOML file *before* we come across the dependency
graph. This CL works around this by doing an extra TOML check when
running in batch mode from within the IDE.

There is a third related problem over on the IDE side, when you run
batch-mode checks but pick a single file. The fix and test is over
there, but this CL also tweaks the MinSdkTooLow lint check to include
the too low level in the error message; that's useful on its own, but
was needed for the IDE-side test related to the scope fix.

Fixes: 410936202
Test: Unit test included
Change-Id: Ic71498e0afe8cc9b01c397510201e8902bdd8dc5