adblib: extract `JdwpProcessPropertiesCollector`
`JdwpProcessPropertiesCollector` is a specific implementation
of collecting `JdwpProcessProperties` for a given `JdwpProcess`.

This change makes this more explicit by changing the
implementation from a "built-in" behavior to a custom extension
provided by the `JdwpProcess.cache`.

This will allow us in the future to provide additional extension
(for example, based only on `trackapp` for API 36+).

Bug: n/a
Test: Included (no visible behavior should change)
Change-Id: If13fd22eb001782edb28963776e32073c1bc228d