adblib: Make coroutine scope cache faster
Make the `CoroutineScopeCache.getOrPutXxx` methods inline so that
they are free of heap allocation (of the `defaultValue` lamdba)
when there is a cache hit, i.e. in the most common cases.

Example:

    val cache: CoroutineScopeCache
    val param: Any
    val value = cache.getOrPut(key) { MyClass(param) }

Before this change the code above would result in
allocating a lamdba at each invocation, even when there
is a cache hit:

    val cache: CoroutineScopeCache
    val param: Any
    val lamdba: () -> MyClass = { MyClass(param) }
    // lamdba is not used when there is a cache hit
    val value = cache.getOrPut(key, lamdba)

After this change, the generated code now looks more like this
(simplified for readability):

    val cache: CoroutineScopeCache
    val param: Any
    // No allocations when calling `getOrDefault`
    val value = cache.getOrDefault(key, null) ?: run {
        // Allocation only if cache miss
        val lamdba: () -> MyClass = { MyClass(param) }
        cache.getOrPut(key, lamdba)
    }

Since `inline` methods are not possible on interfaces, the
`CoroutineScopeCache` is now an abstract class with
`public` inline functions and internal "worker" functions.

Overall, after this change, all `getOrPutXxx` methods are now
allocation free and lock free when there is a "cache hit".

Bug: n/a
Test: Existing
Change-Id: I2cdd69afa5bdca644b4d9663377ad5e071519313