adblib: Introduce `OptionalValue` for `JdwpProcessProperties`
The various values stored in `JdwpProcessProperties` have 3
distinct states:

* The value is not known yet, because the underlying discovery
mechanism is collecting information from the process on the
device

* The value will never be known, due to an error or because
the device may not support retrieving that information from
the process (e.g. `package name` from older APIs)

* The value is known and not-null

So far, we have used `null` to represent the first two states,
but that creates ambiguity for callers, which have to rely
on the `completed` property to know what `null`
actually means.

With this change, we introduce a new `OptionalValue` container
that can either be 1) empty (i.e. uninitialized) or 2) an error
or 3) a value.

The goal is to eventually remove `completed` and `exception` from
`JdwpProcessProperties`.

Bug: n/a
Test: Included
Change-Id: I9f4faca132bfa938851dee012d0abe9eadac05ad