abdlib: remove `completed` from `JdwpProcessProperties`
`JdwpProcessProperties.completed` leaked an implementation detail
that became irrelevant once `trackapp` has become the source of
data for populating `JdwpProcessProperties`. There is no notion
of `completed` with `trackapp` as properties are constantly
watched. Only the "older" implementation, based in `trackjdwp` and
opening `JdwpSession` had a `completed` concept, since it has
to close the `JdwpSession` after a certain timeout.

This change removes the `completed` property, and update the
callers to use `OptionalValue` instead.

Bug: n/a
Test: Updated
Change-Id: I61bbafac9508170219a591e68fe55935790418a0